split bundle module and tests
